Logging and print refactor event module now generic.
Some messages are now suppressed and can be enabled by running
the receiver with the --debug flag.

+++ ap2/connections/event.py
@@ -2,11 +2,13 @@ import socket
 import multiprocessing
 import os
 
-from ..utils import get_logger, get_free_port
+from ..utils import get_file_logger, get_free_port
 
 
-class Event:
-    def __init__(self, addr=None, port=None):
+class EventGeneric:
+    def __init__(self, addr=None, port=None, name='events', isDebug=False):
+        self.name = name
+        self.isDebug = isDebug
         if port is None:
             self.port = get_free_port()
         else:
@@ -15,10 +17,11 @@ class Event:
             self.addr, _ = addr
         else:
             self.addr  = "0.0.0.0"
-        self.file = "./events.bin"
+        self.file = f"./{self.name}.bin"
 
     def serve(self):
-        self.logger = get_logger("event", level="DEBUG")
+        if self.isDebug:
+            self.logger = get_file_logger(self.name, level="DEBUG")
         sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         addr = (self.addr, self.port)
         sock.bind(addr)
@@ -26,8 +29,9 @@ class Event:
 
         try:
             conn, addr = sock.accept()
-            self.logger.debug("Connection open from %s:%d" % addr)
-            event_file = open(self.file, "wb")
+            if self.isDebug:
+                self.logger.debug(f"Open connection from {addr[0]}:{addr[1]}")
+                event_file = open(self.file, "wb")
             try:
                 while True:
                     data = conn.recv(1)
@@ -38,11 +42,13 @@ class Event:
                 pass
             finally:
                 try:
-                    os.remove(self.file)
+                    if self.isDebug:
+                        os.remove(self.file)
                 except OSError:
                     pass
                 conn.close()
-                self.logger.debug("Connection close from %s:%d" % addr)
+                if self.isDebug:
+                    self.logger.debug(f"Close connection to {addr[0]}:{addr[1]}")
             sock.close()
         except KeyboardInterrupt:
             pass
@@ -50,11 +56,17 @@ class Event:
             pass
         finally:
             conn.close()
-            self.socket.close()
+            sock.close()
+            if self.isDebug:
+                self.logger.debug(f"Close connection to {addr[0]}:{addr[1]}")
+
+    # Note that exit handlers and finally clauses, etc., will not be executed.
+    # def terminate(self):
+    #     self.logger.debug(f"Close connection to {addr[0]}:{addr[1]}")
 
     @staticmethod
-    def spawn(addr=None, port=None):
-        event = Event(addr, port)
+    def spawn(addr=None, port=None, name='events', isDebug=False):
+        event = EventGeneric(addr, port, name, isDebug)
         p = multiprocessing.Process(target=event.serve)
         p.start()
         return event.port, p
